+++ cpp-games/src/Player.cpp
@@ -13,14 +13,14 @@ Player::Player(Params &p_in) {
     p = p_in;
 }
 
-void Player::gain(uint32_t gain) {
+void Player::gain(int64_t gain) {
     assert(gain >= 0);
     assets += gain;
     assert(assets > 0); // Players do not exist in a state of long-term debt. 
     // alive = true;       // Players do not exist in a state of long-term debt. 
 }
 
-void Player::lose(uint32_t loss) {
+void Player::lose(int64_t loss) {
     assert(loss >= 0);
     assert(loss <= assets);
     assets -= loss;
@@ -29,7 +29,7 @@ void Player::lose(uint32_t loss) {
     // }
 }
 
-uint32_t Player::get_assets() {
+int64_t Player::get_assets() {
     return assets;
 }
